delete bootstrap functions
